+++ app/ExternalFiles/Extensions/Availability.php
@@ -179,12 +179,11 @@ class Availability extends Extension_Base {
 				<span id="eml_url_file_state"><span class="dashicons dashicons-yes-alt"></span> <?php echo esc_html__( 'File-URL is available.', 'external-files-in-media-library' ); ?></span>
 				<?php
 			} else {
-				$log_url = Helper::get_log_url();
 				?>
 				<span id="eml_url_file_state"><span class="dashicons dashicons-no-alt"></span>
 					<?php
 					/* translators: %1$s will be replaced by the URL for the logs */
-					echo wp_kses_post( sprintf( __( 'File-URL is NOT available! Check <a href="%1$s">the log</a> for details.', 'external-files-in-media-library' ), esc_url( $log_url ) ) );
+					echo wp_kses_post( sprintf( __( 'File-URL is NOT available! Check <a href="%1$s">the log</a> for details.', 'external-files-in-media-library' ), esc_url( Helper::get_log_url( $external_file_obj->get_url( true ) ) ) ) );
 					?>
 					</span>
 				<?php
@@ -259,7 +258,7 @@ class Availability extends Extension_Base {
 		$result = array(
 			'state'   => 'error',
 			/* translators: %1$s will be replaced by the URL for the logs */
-			'message' => sprintf( __( 'URL-File is NOT available! Check <a href="%1$s">the log</a> for details.', 'external-files-in-media-library' ), Helper::get_log_url() ),
+			'message' => sprintf( __( 'File-URL is NOT available! Check <a href="%1$s">the log</a> for details.', 'external-files-in-media-library' ), Helper::get_log_url( $external_file_obj->get_url( true ) ) ),
 		);
 
 		// send response as JSON.