+++ changelog.md
@@ -48,7 +48,7 @@
 - Added option to import real files during synchronization (they are just imported if they are no duplicate)
 - Added privacy hint as checkbox in every import dialog, configurable in user settings
 - Added WordPress Importer entry
-- Added info in admin footer for pages provided by the plugin or for which it makes extensions
+- Added info in admin footer for pages provided by the plugin or for which it makes extensions on the called pages
 - Added support for plugin "WP Extra File Types" to enabled additional possible file types to use as external files
 - Added option to load upload directory via local service
 - Added option to use our plugin name in each HTTP-header User Agent (default enabled)
@@ -60,6 +60,8 @@
 - Added option to show what will be done in import dialog
 - Added support for filenames in other writing systems (like Farsi)
 - Added option to hide the review begging
+- Added new extension to allow import and export of external files in JSON-format (default disabled)
+- Added export and import of settings for external sources
 
 ### Changed